test: improve wait for reported xlogpos
Remove sleep and instead, in WaitClusterSyncedXLogPos, check that the reported
ones are greater or equal a provided XLogPos.

In the various tests get the primary (for standby cluster) or master XLogPos and
use it in the call to WaitClusterSyncedXLogPos.

In TestTimelineFork wait for standbys to have reported their xlogpos just before
stopping them.